requirements.txt
Flask
pandas
openpyxl
gunicorn
Pillow
psycopg2-binary
